reflexion-agent/pyproject.toml
[project]
name = "reflexion-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-tavily>=0.2.15",
    "langgraph>=0.1.0",
    "python-dotenv>=1.0.0",
]
